final release 2

+++ app/src/main/java/com/surelabsid/lti/penilaiankaryawan/response/ResponsePenilaian.kt
@@ -68,5 +68,5 @@ data class PairNilaiXidPointItem(
 	val idPoint: String? = null,
 
 	@field:SerializedName("giveScore")
-	val giveScore: Int? = null
+	val giveScore: Double? = null
 ) : Parcelable
